azure_model/pyproject.toml
[tool.poetry]
name = "azure-crewai-example"
version = "0.1.0"
description = ""
authors = ["ITLackey <itlackey@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
crewai = "0.1.24"
python-dotenv = "1.0.0"
langchain_openai = "0.0.2"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"